backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
sqlmodel
python-dotenv
psycopg2-binary
json-log-formatter
prometheus-fastapi-instrumentator
